Move Stat class so it's nested within Item
