+++ runparser/models/item.py
@@ -15,8 +15,6 @@ import logging
 from ..utils.enums import ItemRarityEnum
 from ..db import RunParserDB
 
-from .stat import Stat
-
 logger = logging.getLogger(__name__)
 
 
@@ -43,7 +41,7 @@ class Item():
 
         self.quote: str = None
         self.desc: str = None
-        self.stats: [Stat] = []
+        self.stats: [self.Stat] = []
         self.unlock: str = None
         self.boss: str = None
 
@@ -65,7 +63,7 @@ class Item():
             if 'desc' in item_data:
                 self.desc = item_data['desc']
             if 'stats' in item_data:
-                self.stats = [Stat(stat_data) for stat_data in item_data['stats']]
+                self.stats = [self.Stat(stat_data) for stat_data in item_data['stats']]
             if 'unlock' in item_data:
                 self.unlock = item_data['unlock']
             if 'boss' in item_data:
@@ -73,6 +71,34 @@ class Item():
 
         except KeyError as e:
             raise self.Item_ParseFailure(f"Item data missing key: {e}")
+
+    class Stat():
+        """Implements the class that quantifies item stats/effects"""
+
+        def __init__(self, data: dict = None):
+            self._reset()
+            if data is not None:
+                self._parse(data)
+
+        def _reset(self):
+            """Resets the Stat object."""
+            self.stat: str = 0
+            self.value = None
+            self.stack: str = None
+            self.stack_add = None
+
+        def _parse(self, data: dict):
+            """Parses the stat data from the given dict."""
+            try:
+                self.stat = data['stat']
+                self.value = data['value']
+                self.stack = data['stack']
+                self.stack_add = data['stack_add']
+            except KeyError as e:
+                raise self.Stat_ParseFailure(f'Missing key: {e}')
+
+        class Stat_ParseFailure(Exception):
+            pass
 
     class Item_ParseFailure(Exception):
         pass
